panning offset now rounds, which fixes some weird aliasing when at 1px/px zooms

+++ src/main/java/com/nettakrim/souper_secret_settings/gui/custom/Panning.java
@@ -118,6 +118,6 @@ public class Panning {
     }
 
     public void setSize(int width, int height) {
-        offset = new Vector2f(width/2f, height/2f);
+        offset = new Vector2f((float)(width/2), (float)(height/2));
     }
 }
